+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/payment_request/payum/command_handlers.xml
@@ -20,8 +20,8 @@
 
         <service id="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\TokenPaymentRequestHandler" abstract="true">
             <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Payum\Resolver\DoctrineProxyObjectResolverInterface" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Payum\Factory\PayumTokenFactoryInterface" />
+            <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
+            <argument type="service" id="sylius.payum.payment_request.factory.payum_token" />
             <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessorInterface" />
             <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\AfterTokenRequestProcessorInterface" />
         </service>
@@ -38,7 +38,7 @@
 
         <service id="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\ModelPaymentRequestHandler" abstract="true">
             <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Payum\Resolver\DoctrineProxyObjectResolverInterface" />
+            <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
             <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessorInterface" />
         </service>
