+++ src/Sylius/Bundle/PayumBundle/PaymentRequest/Factory/PayumTokenFactory.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Payum\Factory;
+namespace Sylius\Bundle\PayumBundle\PaymentRequest\Factory;
 
 use Payum\Core\Payum;
 use Payum\Core\Security\TokenInterface;
