+++ src/Sylius/Bundle/PayumBundle/PaymentRequest/Resolver/DoctrineProxyObjectResolver.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Payum\Resolver;
+namespace Sylius\Bundle\PayumBundle\PaymentRequest\Resolver;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\Proxy;
