+++ src/Sylius/Bundle/PayumBundle/PaymentRequest/Resolver/DoctrineProxyObjectResolverInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Payum\Resolver;
+namespace Sylius\Bundle\PayumBundle\PaymentRequest\Resolver;
 
 use Sylius\Component\Payment\Model\PaymentRequestInterface;
