+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/payment_request/action.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+    xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
+    <services>
+        <defaults public="true" />
+
+        <service id="sylius.payum_action.payment_request.action.get_http_request" class="Sylius\Bundle\PayumBundle\PaymentRequest\Action\SyliusGetHttpRequestAction">
+            <argument type="service" id="sylius.payum.context.payment_request" />
+
+            <tag name="payum.action" all="true" prepend="true" />
+        </service>
+
+        <service id="sylius.payum_action.payment_request.action.render_template" class="Sylius\Bundle\PayumBundle\PaymentRequest\Action\SyliusRenderTemplateAction">
+            <argument type="service" id="sylius.payum.context.payment_request" />
+
+            <tag name="payum.action" all="true" prepend="true" />
+        </service>
+    </services>
+</container>
