+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/payment_request/resolver.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+    xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
+    <services>
+        <defaults public="true" />
+
+        <service id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" class="Sylius\Bundle\PayumBundle\PaymentRequest\Resolver\DoctrineProxyObjectResolver">
+            <argument type="service" id="doctrine.orm.entity_manager" />
+        </service>
+        <service id="Sylius\Bundle\PayumBundle\PaymentRequest\Resolver\DoctrineProxyObjectResolverInterface" alias="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
+    </services>
+</container>
