+++ src/Sylius/Bundle/PayumBundle/composer.json
@@ -1,7 +1,7 @@
 {
     "name": "sylius/payum-bundle",
     "type": "symfony-bundle",
-    "description": "Payum integration for Symfony e-commerce applications.",
+    "description": "PaymentRequest integration for Symfony e-commerce applications.",
     "keywords": [
         "payments",
         "payment",
